New abstract handler for date fields, which the various date handlers extend
