+++ src/Drupal/Driver/Fields/Drupal7/DateHandler.php
@@ -5,5 +5,13 @@ namespace Drupal\Driver\Fields\Drupal7;
 /**
  * Date field handler for Drupal 7.
  */
-class DateHandler extends DatetimeHandler {
+class DateHandler extends AbstractDateHandler {
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function getDateFormat() {
+    return 'Y-m-d\Th:m:s';
+  }
+
 }
