+++ src/Drupal/Driver/Fields/Drupal7/DatestampHandler.php
@@ -5,5 +5,13 @@ namespace Drupal\Driver\Fields\Drupal7;
 /**
  * Datestamp field handler for Drupal 7.
  */
-class DatestampHandler extends DatetimeHandler {
+class DatestampHandler extends AbstractDateHandler {
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function getDateFormat() {
+    return 'U';
+  }
+
 }
